fix: changed the logic of client deletion, removed route including id

+++ src/client/client.service.ts
@@ -1,6 +1,5 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateClientDto } from './dto/create-client.dto';
-import { UpdateClientDto } from './dto/update-client.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { hash } from 'bcrypt';
 
@@ -48,12 +47,20 @@ export class ClientService
     return await this.prisma.client.findUnique( { where: { email: email } } );
   }
 
-  async remove( id: number ) 
+  async delete( id: number ) 
   {
+
+    const client = await this.prisma.client.findUnique( { where: { id: id } } );
+
+    if ( !client )
+    {
+      throw new NotFoundException( "Client doens't exists" );
+    }
+
     return await this.prisma.client.delete(
       {
         where: { id: id }
       }
-    ).catch( () => { throw new NotFoundException( `Can't find item with id ${id} in database` ) } )
+    )
   }
 }
